Removed an unused import and fixed a bug that would cause the bot to tag wrong people sometimes.
tweepy wasn't used in this module so I removed the unused import, a good slight micro-optimization
The checker function was improperly concatenating the username to reply to at the end of the tweet instead of the beginning which caused it to be sometimes cut off (or omitted completely) when the tweet was longer than 121 characters. Removing part of a twitter name changes who is mentioned which can have some bad consequences when tweeting intentionally hateful material.
The checker function now concatenates the username to the front of the tweet causing the name to never be cutoff.

Issues that still exist:
If someone's twitter name is tagged at the end of the tweet it can still get cut off