Moved TweetStreamListener to its own module
For some reason I decided to cleanup the code so that when I merge this
branch into master it will be way better.

Planned stuff:
-Multithreading
---The listener module will handle the replying to tweets
---The bullyranger module will handle telling people to not bully

-Implementing the features
--Bullyranger features

This feature can benefit from the changes to be made in the
nlp_integration branch so I'll probably just delay development on that
until this branch is complete.